+++ utils/hints_test.go
@@ -51,7 +51,7 @@ func TestGetProcessors(t *testing.T) {
 
 func TestGenerateHints(t *testing.T) {
 
-	var allSupportedHints = []string{"enabled", "module", "integration", "datas_treams", "host", "period", "timeout", "metrics_path", "username", "password", "stream", "processors", "multiline", "json", "disable"}
+	var allSupportedHints = []string{"enabled", "package", "module", "integration", "data_streams", "metricsets", "host", "period", "timeout", "metrics_path", "username", "password", "stream", "processors", "multiline", "json", "disable"}
 
 	tests := []struct {
 		name                   string
@@ -117,7 +117,7 @@ func TestGenerateHints(t *testing.T) {
 				"co.elastic.metrics/password":       "pass",
 				"co.elastic.metrics.foobar/period":  "15s",
 				"co.elastic.metrics.foobar1/period": "15s",
-				"co.elastic.hints/steam":            "stdout", // On purpose this added with typo
+				"co.elastic.hints/streamssssssssss": "stdout", // On purpose this added with typo
 				"not.to.include":                    "true",
 			},
 			result: mapstr.M{
@@ -126,7 +126,7 @@ func TestGenerateHints(t *testing.T) {
 						"pattern": "^test",
 					},
 				},
-				"hints": mapstr.M{"steam": "stdout"},
+				"hints": mapstr.M{"streamssssssssss": "stdout"},
 				"metrics": mapstr.M{
 					"module":       "prometheus",
 					"period":       "15s",
@@ -135,7 +135,7 @@ func TestGenerateHints(t *testing.T) {
 					"password":     "pass",
 				},
 			},
-			expectedIncorrectHints: 1, // Due to co.elastic.hints/steam and not co.elastic.hints/stream
+			expectedIncorrectHints: 1, // Due to co.elastic.hints/streamsteamssssssssss
 		},
 		// Scenarios being tested:
 		// logs/multiline.pattern must be a nested mapstr.M under hints.logs
@@ -227,6 +227,79 @@ func TestGenerateHints(t *testing.T) {
 			},
 			expectedIncorrectHints: 0,
 		},
+		// Scenarios being tested:
+		// have co.elastic.hints/package set.
+		// Define multiple co.elastic.hints/data_streams and also specific configuration for each one
+		// Typo errors introduced for "co.elastic.hints/access.streams" and "co.elastic.hints/error.streams"
+		{
+			name: "Metrics_apache_package_and_specific_config_per_datastream",
+			annotations: map[string]string{
+				"co.elastic.hints/package":                 "apache",
+				"co.elastic.hints/data_streams":            "access,error",
+				"co.elastic.hints/access.period":           "5m",
+				"co.elastic.hints/access.streamssssssssss": "stdout", // On purpose this added with typo
+				"co.elastic.hints/error.period":            "5m",
+				"co.elastic.hints/error.streamssssssssss":  "stderr", // On purpose this added with typo
+			},
+			result: mapstr.M{
+				"hints": mapstr.M{
+					"data_streams": "access,error",
+					"access":       mapstr.M{"period": "5m", "streamssssssssss": "stdout"},
+					"error":        mapstr.M{"period": "5m", "streamssssssssss": "stderr"},
+					"package":      "apache",
+				}},
+			expectedIncorrectHints: 2, // Due to co.elastic.hints/access.streamssssssssss and co.elastic.hints/error.streamssssssssss typo errors
+		},
+		// Scenarios being tested:
+		// have co.elastic.metrics/module set.
+		// Define multiple co.elastic.hints/data_streams and also specific configuration for each one
+		// A typo error introduced for "co.elastic.metrics/istiod.streams"
+		{
+			name: "Metrics_istio_module_and_specific_config_per_metricset",
+			annotations: map[string]string{
+				"co.elastic.metrics/module":                  "istio",
+				"co.elastic.metrics/metricsets":              "istiod,proxy",
+				"co.elastic.metrics/istiod.period":           "5m",
+				"co.elastic.metrics/istiod.streamssssssssss": "stdout", // On purpose this added with typo
+				"co.elastic.metrics/proxy.period":            "5m",
+				"co.elastic.metrics/proxy.stream":            "stderr",
+			},
+			result: mapstr.M{
+				"metrics": mapstr.M{
+					"metricsets": "istiod,proxy",
+					"istiod":     mapstr.M{"period": "5m", "streamssssssssss": "stdout"},
+					"proxy":      mapstr.M{"period": "5m", "stream": "stderr"},
+					"module":     "istio",
+				}},
+			expectedIncorrectHints: 1, // Due to co.elastic.metrics/istiod.streamssssssssss
+		},
+		// Scenarios being tested:
+		// have co.elastic.metrics/module set for multiple enumerations.
+		// Define different hints for each one enumeration
+		// A typo error introduced for "co.elastic.metrics/1.periods" and "co.elastic.metrics/2.streams"
+		{
+			name: "Metrics_multiple_modules_and_specific_config_per_module",
+			annotations: map[string]string{
+				"co.elastic.metrics/1.module":           "prometheus",
+				"co.elastic.metrics/1.periodssssssssss": "15s", // On purpose this added with typo
+				"co.elastic.metrics/2.module":           "istiod",
+				"co.elastic.metrics/2.period":           "15s",
+				"co.elastic.metrics/2.streamssssssssss": "stderr", // On purpose this added with typo
+			},
+			result: mapstr.M{
+				"metrics": mapstr.M{
+					"1": mapstr.M{
+						"module":           "prometheus",
+						"periodssssssssss": "15s",
+					},
+					"2": mapstr.M{
+						"module":           "istiod",
+						"period":           "15s",
+						"streamssssssssss": "stderr",
+					},
+				}},
+			expectedIncorrectHints: 2, // Due to co.elastic.metrics/1.periodssssssssss and co.elastic.metrics/2.streamssssssssss typo errors
+		},
 	}
 
 	for _, test := range tests {
@@ -237,6 +310,7 @@ func TestGenerateHints(t *testing.T) {
 				continue
 			}
 		}
+
 		generateHints, incorrectHints := GenerateHints(annMap, "foobar", "co.elastic", allSupportedHints)
 		assert.Equal(t, test.expectedIncorrectHints, len(incorrectHints)) // We validate how many incorrect hints are provided per test case.
 		assert.Equal(t, test.result, generateHints)