+++ CHANGELOG.md
@@ -53,3 +53,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 
 [0.6.9]: https://github.com/elastic/elastic-agent-autodiscover/compare/v0.6.8...v0.6.9
+
+## [0.6.11]
+
+### Changed
+
+- Enhance GenerateHints function to check supported list of hints for multiple datastreams and metricsets
+
+
+[0.6.10]: https://github.com/elastic/elastic-agent-autodiscover/compare/v0.6.10...v0.6.11